refactor(server): packagelize dependencies
